Updated budget from integer to double

+++ code/back-end/src/main/java/met/cs673/team1/domain/dto/UserGetDto.java
@@ -29,6 +29,6 @@ public class UserGetDto {
     private List<Role> roles;
 
     @JsonProperty
-    private Integer budget;
+    private Double budget;
 
 }
