+++ code/back-end/src/main/java/met/cs673/team1/domain/dto/UserPostDto.java
@@ -35,5 +35,5 @@ public class UserPostDto {
 
     @JsonProperty
     @NotEmpty
-    private Integer budget;
+    private Double budget;
 }
